refactor: make path more generic
Currently the server expects a `PUT` request with a payload defining the
query and relation to be sent to the `/analyse` endpoint. This is hard
to add to, for example to return all locks taken regardless of relation,
so we should make it more generic and use better REST semantics.

This change:
* Updates the endpoint to `/locks/:relation`
* Changes it to be a `GET` method
* Updates the server and CLI to handle this