feat: various changes
Few different changes in here, based on testing at work. Firstly, it's
useful to be able to specify the password to connect to the database as
it won't always be a password-less user.

Secondly, even for a single relation there may be multiple locks taken.
This is true in the case of foreign keys at least, which take an
`AccessShareLock` as well as `ShareRowExclusiveLock` on the parent table
according to `pg_locks`.

Finally, typing out the query into the terminal isn't great fun
especially if it's long. Copy-pasting also doesn't seem to work. `curl`
supports reading from a file with the `@` prefix, so let's support that
too.

This change:
* Adds some new features and usability improvements