feat: show better error messages
When the server returns a 500 (usually because the SQL provided is
invalid), neither it nor the client display any messages about what just
happened (the CLI just displays a generic error message). This is
because `ureq` hasn't read the response body yet necessarily, so it just
says it got back a 500.

Instead, we can format the error messages a little more nicely
ourselves.

This change:
* Adds some handling to `make_request` to provide better error messages
  from the CLI