refactor: house a single binary
It's not a great user experience to have to run `cargo run --bin ...`
all the time and it means we're running debug binaries all over the
place. It also makes the examples more difficult to read and feels less
polished.

Instead, we can move all the code under a single binary (now called
`pglx`) that has 2 subcommands:

* `pglx serve` runs the server component
* `pglx query` runs the CLI component

This allows us to provide better help messages and argument references
as we're using `clap` over `pico-args` for a better experience.

This change:
* Restructures all the code back into a single binary with subcommands