feat: support more lock parsing options
`AccessExclusiveLock` is what PostgreSQL returns internally when making
queries in `pg_locks`, so we need to support that as a priority
otherwise we cannot parse the query results. However, explicit locking
usually takes the form of:

```sql
LOCK TABLE foo IN ACCESS EXCLUSIVE MODE;
```

so it would be worth supporting this. The `Lock` suffix is also a little
annoying to remember so we can support `AccessExclusive` as well.

This change:
* Updates the `from_str` method to support some other forms